+++ CHANGELOG.md
@@ -2,7 +2,8 @@ CHANGELOG for Sulu
 ==================
 
 * dev-master
-    * HOTFIX      #1568 [ContentBundle]  Fixed copy extension data
+    * HOTFIX      #1568 [ContentBundle]         Fixed copy extension data
+    * HOTFIX      #1577 [DocumentManagerBundle] Made caching directory for DocumentManager configureable
 
 * 1.0.8 (2015-08-31)
     * HOTFIX      #1539 [WebsiteBundle]  Fixed canonical tag for shadow pages